requirements.txt
Django==3.2.12
python-decouple==3.6
psycopg2-binary==2.9.3
django-storages==1.12.3
boto3==1.21.16
django-grappelli==3.0.3
django-extensions==3.1.5
shortuuid==1.0.8
django-ckeditor==6.2.0
django-ipware==4.0.2
channels==3.0.4
channels-redis==3.4.0
django-redis==5.2.0
selenium==4.1.3
djangorestframework==3.13.1
serpy==0.3.1
drf-flex-fields==0.9.7
djoser==2.1.0